+++ services/madoc-ts/src/frontend/shared/capture-models/editor/input-types/CheckboxListField/index.ts
@@ -16,6 +16,7 @@ const specification: FieldSpecification<CheckboxListFieldProps> = {
   label: 'Checkbox list field',
   defaultProps: {},
   allowMultiple: true,
+  maxMultiple: 99,
   required: false,
   TextPreview: CheckboxListFieldPreview,
   description: 'List of checkboxes',