+++ services/madoc-ts/src/frontend/shared/capture-models/editor/input-types/TaggedTextField/index.ts
@@ -17,6 +17,7 @@ const specification: FieldSpecification<TaggedTextFieldProps> = {
   Component: React.lazy(() => import(/* webpackChunkName: "fields" */ './TaggedTextField')),
   defaultValue: '',
   allowMultiple: true,
+  maxMultiple: 99,
   required: false,
   defaultProps: {
     preset: 'bentham',